DEV Create s/o/detectors.py with Detector object
This module will be used to create all the subclassed
detector instances for mesh detectors
