+++ serpentTools/objects/__init__.py
@@ -1,6 +1,6 @@
 """Objects used to support the parsing."""
 
-from serpentTools.objects.containers import (
-        HomogUniv, Detector, BranchContainer)
+from serpentTools.objects.containers import HomogUniv, BranchContainer
+from serpentTools.objects.detectors import Detector
